implement auth using gates

+++ app/Http/Controllers/QuestionsController.php
@@ -60,10 +60,13 @@ class QuestionsController extends Controller
      * @param  \App\Question  $questions
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit(Question $question)
     {
-        $question = Question::findOrFail($id);
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, 'access denied!');
+        }
         return view('questions.edit', compact('question'));
+
     }
 
     /**
@@ -75,6 +78,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionResquest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, 'access deniedd!');
+        }
         $question->update($request->only('title', 'body'));
         session()->flash('success', 'question updated successfuly');
         return redirect('/questions')->with('success', 'your question has been updated');
@@ -88,6 +94,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)){
+            session()->flash('error', 'access denied');
+            return redirect()->back();
+        }
         $question->delete();
         session()->flash('success', 'question has been deleted successfuly');
         return redirect('/questions');
